General Ajax
1. 增加了游戏页面外ajax的功能
2. 修正了pdo_mysql类的一个BUG
3. 增加了gametype参数

+++ gamedata/mod/dtsp/sql/players.dtsp.sql
@@ -29,7 +29,7 @@
   baseatt smallint unsigned NOT NULL default '0',
   basedef smallint unsigned NOT NULL default '0',
   region tinyint unsigned NOT NULL default '0',
-  area tinyint unsigned NOT NULL default '0',
+  area smallint unsigned NOT NULL default '0',
   lvl tinyint unsigned NOT NULL default '0',
   `exp` smallint unsigned NOT NULL default '0',
   `upexp` smallint unsigned NOT NULL default '0',
